+++ .gitignore
@@ -106,3 +106,7 @@ dist
 .tern-port
 
 package-lock.json
+
+# build output
+src/static/
+dist/
