--- assets/js/main.js
@@ -1,1 +1,0 @@
-import "../scss/styles.scss";
