--- assets/scss/styles.scss
@@ -1,3 +1,0 @@
-body{
-    background-color: powderblue;
-}
