--- controllers/mainController.js
@@ -1,1 +1,0 @@
-export const home = (req, res) => res.render("home");
