--- globalRouter.js
@@ -1,9 +1,0 @@
-import express from "express";
-import routes from "./routes";
-import { home } from "./controllers/mainController";
-
-const globalRouter = express.Router(); 
-
-globalRouter.get(routes.home, home);
-
-export default globalRouter;
