--- init.js
@@ -1,9 +1,0 @@
-import app from "./app";
-
-const PORT = 4000;
-
-const handleListening = () => {
-    console.log(`✅ Listening on: http://localhost:${PORT}`)
-}
-
-app.listen(PORT, handleListening)
