+++ package.json
@@ -2,12 +2,7 @@
   "name": "save-idiot-like-someone",
   "version": "1.0.0",
   "description": "mentoring website",
-  "main": "index.js",
-  "scripts": {
-    "dev:server": "nodemon --exec babel-node init.js -- delay 2",
-    "dev:assets": "WEBPACK_ENV=development webpack -w",
-    "build:assets": "WEBPACK_ENV=production webpack"
-  },
+  "main": "main.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/HINCS/save-idiot-like-someone.git"
@@ -32,12 +27,23 @@
     "pug": "^2.0.4",
     "sass-loader": "^8.0.0",
     "webpack": "^4.41.2",
-    "webpack-cli": "^3.3.10"
+    "webpack-cli": "^3.3.10",
+    "dotenv": "^8.2.0"
   },
   "devDependencies": {
     "eslint": "^6.7.0",
     "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-import": "^2.18.2",
-    "nodemon": "^2.0.1"
+    "eslint-plugin-prettier": "^3.1.1",
+    "nodemon": "^2.0.1",
+    "prettier": "^1.19.1"
+  },
+  "scripts": {
+    "dev:server": "nodemon --exec babel-node src/init.js -- delay 2",
+    "dev:assets": "cd src && WEBPACK_ENV=development webpack -w",
+    "build:assets": "cd src && WEBPACK_ENV=production webpack",
+    "lint": "eslint src/",
+    "fix": "eslint --fix src/"
   }
 }
