--- routes.js
@@ -1,8 +1,0 @@
-//Global
-const HOME = "/";
-
-const routes = {
-    home: HOME
-}
-
-export default routes;
