+++ src/assets/js/main.js
@@ -1,0 +1,1 @@
+import "../scss/styles.scss";
