+++ src/assets/scss/styles.scss
@@ -1,0 +1,3 @@
+body {
+  background-color: powderblue;
+}
