+++ src/controllers/introController.js
@@ -1,0 +1,2 @@
+const intro = (req, res) => res.render("intro");
+export default intro;
