+++ src/init.js
@@ -1,0 +1,12 @@
+import * as dotenv from "dotenv";
+import app from "./app";
+
+dotenv.config();
+
+const PORT = process.env.PORT || 4000;
+
+const handleListening = () => {
+  console.log(`✅ Listening on: http://localhost:${PORT}`);
+};
+
+app.listen(PORT, handleListening);
