+++ src/routes.js
@@ -1,0 +1,8 @@
+// Global Routes
+const INTRO = "/";
+
+const routes = {
+  intro: INTRO
+};
+
+export default routes;
