+++ src/views/intro.pug
@@ -1,0 +1,8 @@
+doctype html
+html
+    head
+        link(rel="stylesheet", href="/static/styles.css")
+    body
+        p Hello Javascript !!!
+        p This is intro page.
+        script(src="/static/main.js")
