--- static/styles.css
@@ -1,2 +1,0 @@
-body {
-  background-color: powderblue; }
