--- views/home.pug
@@ -1,7 +1,0 @@
-doctype html
-html
-    head
-        link(rel="stylesheet", href="/static/styles.css")
-    body
-        script(src="/static/main.js")
-        p Hi!
